add scripted pipeline

+++ Jenkinsfile
@@ -1,52 +1,34 @@
-pipeline {
-    agent none
-    options {
-        skipStagesAfterUnstable()
+node {
+    skipStagesAfterUnstable()
+    docker.image('python:2-alpine').inside() {
+        stage('Build') { 
+            sh 'python -m py_compile sources/add2vals.py sources/calc.py' 
+            stash(name: 'compiled-results', includes: 'sources/*.py*')
+        }
     }
-    stages {
-        stage('Build') {
-            agent {
-                docker {
-                    image 'python:2-alpine'
-                }
-            }
-            steps {
-                sh 'python -m py_compile sources/add2vals.py sources/calc.py'
-                stash(name: 'compiled-results', includes: 'sources/*.py*')
-            }
+
+    docker.image('qnib/pytest').inside() {
+        stage('Test') { 
+            sh 'py.test --verbose --junit-xml test-reports/results.xml sources/test_calc.py' 
+            junit 'test-reports/results.xml'
         }
-        stage('Test') {
-            agent {
-                docker {
-                    image 'qnib/pytest'
-                }
-            }
-            steps {
-                sh 'py.test --junit-xml test-reports/results.xml sources/test_calc.py'
-            }
-            post {
-                always {
-                    junit 'test-reports/results.xml'
-                }
-            }
-        }
+    }
+
+    docker.image('cdrx/pyinstaller-linux:python2').inside() {
+        withEnv(
+            ['VOLUME="$(pwd)/sources:/src"',
+             'IMAGE=cdrx/pyinstaller-linux:python2'
+            ]
+        )
         stage('Deliver') { 
-            agent any
-            environment { 
-                VOLUME = '$(pwd)/sources:/src'
-                IMAGE = 'cdrx/pyinstaller-linux:python2'
-            }
-            steps {
-                dir(path: env.BUILD_ID) { 
-                    unstash(name: 'compiled-results') 
-                    sh "docker run --rm -v ${VOLUME} ${IMAGE} 'pyinstaller -F add2vals.py'" 
-                }
-            }
-            post {
-                success {
-                    archiveArtifacts "${env.BUILD_ID}/sources/dist/add2vals" 
-                    sh "docker run --rm -v ${VOLUME} ${IMAGE} 'rm -rf build dist'"
-                }
+            dir(path: env.BUILD_ID) { 
+                unstash(name: 'compiled-results') 
+                sh "docker run --rm -v ${VOLUME} ${IMAGE} 'pyinstaller -F add2vals.py'" 
+            } 
+            def currentResult = currentBuild.result ?: 'SUCCESS'
+            if (currentResult == 'SUCCESS') {
+                archiveArtifacts "${env.BUILD_ID}/sources/dist/add2vals" 
+                sh "docker run --rm -v ${VOLUME} ${IMAGE} 'rm -rf build dist'"
             }
         }
     }
